+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Option to disable browser feature detection (#3)
+
 ## [1.1.0] - 2021-10-13
 
 ### Added